Changed db variables for import only (no tagging).
I upped the size of datatypes for the table used when sequence reads are imported to mysql.  Largely, this was a result of trying to run S. mansoni contigs through the program, which tend to be larger than your average 454 read.  Essentially changed all SMALLINT to MEDIUMINT and BLOB to MEDIUMBLOB and TEXT to MEDIUMTEXT.