Changing gap char for linker fuzzy matching
The new linkers tend to have a fair amount of sequence at their 5' ends.  So, rather than match against this sequence, which is not terribly unique, allow the fuzzy match between the more 3' elements of the linker sequence which *are* unique.

Also, bump up dbase values to hold longer linker sequences.